[Store] Move Distance classes to dedicated namespace
